+++ source/Core/Smarty/Plugin/EmosAdapter.php
@@ -46,6 +46,8 @@ use OxidEsales\Eshop\Core\Str;
  * This class is a reference implementation of a PHP Function to include
  * ECONDA Trackiong into a Shop-System.
  *
+ * @deprecated since v6.5.6 (2020-07-29); moved the functionality to 'OXID personalization' module
+ *
  * The smarty tempaltes should include s tag like
  * [{insert name="oxid_tracker" title=$template_title}]
  */