+++ tests/Codeception/acceptance/DownloadableProductCest.php
@@ -9,6 +9,7 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\CodeceptionAdmin;
 
+use DateTime;
 use OxidEsales\Codeception\Module\Translation\Translator;
 use OxidEsales\Codeception\Page\Home;
 use OxidEsales\Codeception\Step\Basket;
@@ -16,6 +17,29 @@ use OxidEsales\EshopCommunity\Tests\Codeception\AcceptanceTester;
 
 final class DownloadableProductCest
 {
+    /**
+     * @var string
+     */
+    private $orderId;
+
+    /** @param AcceptanceTester $I */
+    public function _before(AcceptanceTester $I)
+    {
+        $I->updateConfigInDatabase('iMaxDownloadsCount', "2", 'str');
+        $I->updateConfigInDatabase('iLinkExpirationTime', "240", 'str');
+        $I->updateInDatabase('oxarticles', ['oxisdownloadable' => 1], ['oxartnum' => '1002-1']);
+    }
+
+    /** @param AcceptanceTester $I */
+    public function _after(AcceptanceTester $I)
+    {
+        $I->updateConfigInDatabase('iMaxDownloadsCount', "0", 'str');
+        $I->updateConfigInDatabase('iLinkExpirationTime', "168", 'str');
+        $I->updateInDatabase('oxarticles', ['oxisdownloadable' => 0], ['oxartnum' => '1002-1']);
+        $I->deleteFromDatabase('oxorder', ['OXID' => $this->orderId]);
+        $I->deleteFromDatabase('oxorderarticles', ['OXORDERID' => $this->orderId]);
+    }
+
     /** @param AcceptanceTester $I */
     public function downloadableFiles(AcceptanceTester $I): void
     {
@@ -23,9 +47,41 @@ final class DownloadableProductCest
 
         $I->clearShopCache();
         $startPage = $I->loginShopWithExistingUser();
+        $this->makePurchaseComplete($I);
 
-        $this->enableDownloadableFilesAndSetForAProduct($I);
-        $this->makePurchaseComplete($I, $startPage);
+        $this->orderId = $I->grabFromDatabase('oxorder', 'OXID', ['oxuserid' => 'testuser']);
+        $articleId = $I->grabFromDatabase('oxorderarticles', 'OXID', ['OXORDERID' => $this->orderId]);
+
+        $I->haveInDatabase(
+            'oxorderfiles',
+            [
+                'OXID' => "testdownloadProductCest",
+                'OXORDERID' => $this->orderId,
+                'OXFILENAME' => 'testFile3',
+                'OXFILEID' => '1000l',
+                'OXSHOPID' => 1,
+                'OXORDERARTICLEID' => $articleId,
+                'OXDOWNLOADCOUNT' => '0',
+                'OXMAXDOWNLOADCOUNT' => 2,
+                'OXDOWNLOADEXPIRATIONTIME' => 24,
+                'OXLINKEXPIRATIONTIME' => 240,
+                'OXRESETCOUNT' => 0,
+                'OXVALIDUNTIL' => (new DateTime())->modify('+1 week')->format('Y-m-d 00:00:00'),
+                'OXTIMESTAMP' => (new DateTime())->format('Y-m-d 00:00:00')
+            ]
+        );
+
+        $I->haveInDatabase(
+            'oxfiles',
+            [
+                'OXID' => '1000l',
+                'OXARTID' => '1002-1',
+                'OXFILENAME' => 'testFile3',
+                'OXPURCHASEDONLY' => 1,
+                'OXSTOREHASH' => 'e48a1b571bd2d2e60fb2d9b1b76b35d5',
+            ]
+        );
+
         $this->checkMyDownloads($I, $startPage);
         $this->makeOrderComplete($I);
         $this->checkFileInMyDownloads($I, $startPage);
@@ -34,22 +90,7 @@ final class DownloadableProductCest
     /**
      * @param AcceptanceTester $I
      */
-    private function enableDownloadableFilesAndSetForAProduct(AcceptanceTester $I): void
-    {
-        $I->updateConfigInDatabase('blEnableDownloads', true, 'bool');
-        $I->updateConfigInDatabase('iMaxDownloadsCount', "2", 'str');
-        $I->updateConfigInDatabase('iLinkExpirationTime', "240", 'str');
-        $I->updateConfigInDatabase('iDownloadExpirationTime', "24", 'str');
-        $I->updateConfigInDatabase('iMaxDownloadsCountUnregistered', "1", 'str');
-
-        $I->updateInDatabase('oxarticles', ['oxisdownloadable' => 1], ['oxartnum' => '1002-1']);
-    }
-
-    /**
-     * @param AcceptanceTester $I
-     * @param Home             $startPage
-     */
-    private function makePurchaseComplete(AcceptanceTester $I, Home $startPage): void
+    private function makePurchaseComplete(AcceptanceTester $I): void
     {
         $basket = new Basket($I);
         $userCheckoutPage = $basket->addProductToBasketAndOpenUserCheckout('1002-1', 1);
@@ -77,7 +118,7 @@ final class DownloadableProductCest
     private function makeOrderComplete(AcceptanceTester $I): void
     {
         $currentTime = date('Y-m-d H:i:s');
-        $I->updateInDatabase('oxorder', ['oxpaid' => $currentTime], ['oxordernr' => 2]);
+        $I->updateInDatabase('oxorder', ['oxpaid' => $currentTime], ['OXID' => $this->orderId]);
     }
 
     /**